pyproject.toml
[project]
name = "xiaohongshu-generator"
version = "0.1.0"
description = "小红书图文生成器"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []
